Groovy: shorten comments and move NegativeCachingGroovyClassLoader to internal package
