+++ rewrite-groovy/src/test/java/org/openrewrite/groovy/GradleSnippetGlobalTransformReproTest.java
@@ -34,16 +34,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 /**
  * Guards against the failure mode behind "Failed to parse build.gradle at cursor position N" in
  * {@code UpgradeTransitiveDependencyVersion} (and openrewrite/rewrite#7870): a <em>global</em> Groovy AST
- * transformation present on the classpath the parser discovers transformations from rewrites the AST away
- * from the source, breaking the position-coupled {@link GroovyParserVisitor}.
+ * transformation on the classpath rewrites the AST away from the source, breaking the position-coupled
+ * {@link GroovyParserVisitor}.
  *
- * <p>{@code GroovyParser} discovers global AST transformations through a <em>classpath-free</em> loader
- * (and disables whatever that loader still finds on the runtime classpath, e.g. groovy-core's {@code @Grab}
- * handler). A transformation registered only on the compile classpath the parser is given is therefore never
- * discovered, so it cannot run — making parsing independent of which transformations happen to be on the
- * compile classpath. This also doubles as a guard that the per-source {@code transformLoader} stays
- * classpath-free: were it to carry the compile classpath again, the transformation below would be discovered,
- * would not be in the disabled set, and would run.
+ * <p>{@code GroovyParser} discovers global AST transformations through a classpath-free loader, so one
+ * registered only on the compile classpath is never discovered and cannot run. This doubles as a guard that
+ * the per-source {@code transformLoader} stays classpath-free: were it to carry the compile classpath again,
+ * the transformation below would run.
  */
 class GradleSnippetGlobalTransformReproTest {
 
@@ -78,8 +75,7 @@ class GradleSnippetGlobalTransformReproTest {
 
         // Before the fix this produced a ParseError: "Failed to parse build.gradle at cursor position N".
         assertThat(parsed).isNotInstanceOf(ParseError.class);
-        // The transformation is registered only on the compile classpath, which the parser's classpath-free
-        // transformLoader does not scan, so it is never discovered and cannot run.
+        // Registered only on the compile classpath, which the classpath-free transformLoader never scans.
         assertThat(InjectClosureStatementTransformation.executed)
                 .as("a global AST transformation registered only on the compile classpath must not run")
                 .isFalse();
